Global daily deaths from COVID-19 prediction
Based on the COVID-19 dataset published by ECDC here:
https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide

Exploratory analysis shows that there is correlation between the global daily deaths and global daily new cases from COVID-19. This is a preliminary effort to fit and predict from a linear model. The large MSE and deviations that reach 30% suggest to try a non-linear model.